pyproject.toml
[tool.poetry]
name = "reserve-ease-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
ruff = "^0.1.6"
sqlalchemy = "^2.0.23"
pandas = "^2.1.3"
supabase = "^2.2.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.2"
email-validator = "^2.1.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = []
ignore = []
fixable = ["ALL"]
target-version = "py310"

# Exclude a variety of commonly ignored directories.
exclude = []